Fikser en feil ved cast-ing. ``` java.lang.ClassCastException: class java.lang.Long cannot be cast to class java.math.BigInteger (java.lang.Long and java.math.BigInteger are in module java.base of loader 'bootstrap') ```

+++ src/main/kotlin/no/nav/cv/notifikasjon/StatusRepository.kt
@@ -123,7 +123,7 @@ open class StatusRepository(
     open fun antallAvStatus(status: String): Long
         = (entityManager.createNativeQuery(antallAvStatusQuery)
             .setParameter("status", status)
-            .singleResult as BigInteger).toLong()
+            .singleResult as Long)
 
 
     private val antallFerdigQuery =
@@ -139,7 +139,7 @@ open class StatusRepository(
             = (entityManager.createNativeQuery(antallFerdigQuery)
                 .setParameter("varsletStatus", varsletStatus)
                 .setParameter("nesteStatus", nesteStatus)
-                .singleResult as BigInteger).toLong()
+                .singleResult as Long)
 
     private val conversionRateWeeklyQuery =
         """
